+++ Albums/Albums/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
