--- Albums/Albums/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Albums
-//
-//  Created by Nonye on 5/7/20.
-//  Copyright © 2020 Nonye Ezekwo. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
